pyproject.toml
[tool.poetry]
name = "kc2re-agent"
version = "0.1.0"
description = ""
authors = ["otariidae <otariidae@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
websocket-client = "^1.1.0"
pylxd = "^2.3.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
mypy = "^0.910"
types-requests = "^2.25.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
